requirements.txt
astroid==2.3.3
autopep8==1.4.4
cycler==0.10.0
decorator==4.4.1
imageio==2.6.1
isort==4.3.21
joblib==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
networkx==2.4
numpy==1.17.4
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.3
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
